requirements.txt
streamlit
pandas
matplotlib
requests
biopython
google-generativeai
seaborn
